[obey-campaign:8deed8b4] test(dungeon): split move integration coverage
